refactor(search-box): extract shared autocomplete segment resolver
